move response and request to their own files
